Now properly dealing with the (rare) case of changed access token
The access token can change if the user logs in with one browser, then logs in again with another browser. Unless the user refreshes the page in the first browser, the access token (held in an AccessToken obj) will be stale.  This change refreshes the access token before making the call to ESDR to create the client.